(refactor) buttons folder structure
