+++ app/src/components/common/index.ts
@@ -3,17 +3,6 @@ export { BalanceShares } from './balance_shares'
 export { BalanceToken } from './balance_token'
 export { BarDiagram } from './bar_diagram_probabilities'
 export { BigNumberInput } from './big_number_input'
-export { Button } from './button'
-export { ButtonAdd } from './button_add'
-export { ButtonAnchor } from './button_anchor'
-export { ButtonCircle } from './button_circle'
-export { ButtonConnectWallet } from './button_connect_wallet'
-export { ButtonContainer } from './button_container'
-export { ButtonDisconnectWallet } from './button_disconnect_wallet'
-export { ButtonLink } from './button_link'
-export { ButtonSelectable } from './button_selectable'
-export { ButtonStateful } from './button_stateful'
-export { ButtonTab } from './button_tab'
 export { Categories } from './categories'
 export { Checkbox } from './checkbox'
 export { ClosedMarket } from './closed_market'
