+++ app/src/components/common/wrong_network_message/index.tsx
@@ -3,8 +3,7 @@ import { withRouter } from 'react-router-dom'
 import styled from 'styled-components'
 
 import { networkIds } from '../../../util/networks'
-import { ButtonContainer } from '../button_container'
-import { ButtonLink } from '../button_link'
+import { ButtonContainer, ButtonLink } from '../../button'
 import { SectionTitle } from '../section_title'
 import { SubsectionTitle } from '../subsection_title'
 import { ViewCard } from '../view_card'
