+++ app/src/components/market/market_home.tsx
@@ -2,9 +2,10 @@ import React, { useCallback, useEffect, useState } from 'react'
 import styled from 'styled-components'
 
 import { ConnectedWeb3Context } from '../../hooks/connectedWeb3'
-import { ButtonType } from '../../theme/component_styles/button_styling_types'
 import { RemoteData } from '../../util/remote_data'
-import { Button, ButtonCircle, ButtonSelectable, ListCard, ListItem, SectionTitle } from '../common'
+import { Button, ButtonCircle, ButtonSelectable } from '../button'
+import { ButtonType } from '../button/button_styling_types'
+import { ListCard, ListItem, SectionTitle } from '../common'
 import { AdvancedFilters } from '../common/advanced_filters'
 import { Dropdown, DropdownItemProps, DropdownPosition } from '../common/dropdown'
 import { IconFilter } from '../common/icons/IconFilter'
