+++ app/src/components/market/profile/closed_market_detail.tsx
@@ -9,15 +9,8 @@ import { CPKService, ERC20Service } from '../../../services'
 import { getLogger } from '../../../util/logger'
 import { formatBigNumber, formatDate } from '../../../util/tools'
 import { Arbitrator, BalanceItem, OutcomeTableValue, Status, Token } from '../../../util/types'
-import {
-  Button,
-  ButtonContainer,
-  ClosedMarket,
-  DisplayArbitrator,
-  SubsectionTitle,
-  TitleValue,
-  ViewCard,
-} from '../../common'
+import { Button, ButtonContainer } from '../../button'
+import { ClosedMarket, DisplayArbitrator, SubsectionTitle, TitleValue, ViewCard } from '../../common'
 import { OutcomeTable } from '../../common/outcome_table'
 import { FullLoading } from '../../loading'
