+++ app/src/components/market/steps/items/summary_market_step.tsx
@@ -5,11 +5,10 @@ import styled from 'styled-components'
 import { MARKET_FEE } from '../../../../common/constants'
 import { useMarketMakerData, useQuestion } from '../../../../hooks'
 import { useConnectedWeb3Context } from '../../../../hooks/connectedWeb3'
-import { ButtonType } from '../../../../theme/component_styles/button_styling_types'
 import { formatBigNumber, formatDate } from '../../../../util/tools'
+import { ButtonAnchor, ButtonContainer } from '../../../button'
+import { ButtonType } from '../../../button/button_styling_types'
 import {
-  ButtonAnchor,
-  ButtonContainer,
   CopyText,
   CreateCard,
   DisplayArbitrator,
