+++ app/src/components/market/transaction_details_card/index.tsx
