+++ app/src/components/market/transaction_details_row/index.tsx
