+++ app/src/components/modal/modal_question/index.tsx
@@ -3,16 +3,12 @@ import styled from 'styled-components'
 
 import { useAsyncDerivedValue, useContracts } from '../../../hooks'
 import { ConnectedWeb3Context } from '../../../hooks/connectedWeb3'
-import { ButtonType } from '../../../theme/component_styles/button_styling_types'
 import { getArbitratorFromAddress } from '../../../util/networks'
 import { formatDate, truncateStringInTheMiddle } from '../../../util/tools'
 import { Arbitrator, Question } from '../../../util/types'
-import { Button } from '../../common/button'
-import { FormRow } from '../../common/form_row'
-import { Spinner } from '../../common/spinner'
-import { SubsectionTitle } from '../../common/subsection_title'
-import { Textfield } from '../../common/textfield'
-import { TitleValue } from '../../common/title_value'
+import { Button } from '../../button'
+import { ButtonType } from '../../button/button_styling_types'
+import { FormRow, Spinner, SubsectionTitle, Textfield, TitleValue } from '../../common'
 import { ModalWrapper } from '../../modal/modal_wrapper'
 
 const ButtonStyled = styled(Button)`
